Move passClasses to archivematicaCommon
These classes have no actual dependencies on any contents of MCPServer,
and are now being used in multiple modules, so move them to
archivematicaCommon.

The module is renamed from passClasses to dicts.

Refs #6451.
