+++ src/MCPServer/lib/linkTaskManagerDirectories.py
@@ -23,7 +23,6 @@
 
 from linkTaskManager import LinkTaskManager
 from taskStandard import taskStandard
-from passClasses import ReplacementDict
 import os
 import sys
 import threading
@@ -32,6 +31,7 @@ sys.path.append("/usr/lib/archivematica/archivematicaCommon")
 import databaseInterface
 import databaseFunctions
 from databaseFunctions import deUnicode
+from dicts import ReplacementDict
 
 
 class linkTaskManagerDirectories(LinkTaskManager):
